requirements.txt
streamlit
google-generativeai
black